Moved `plugins` mod to its own file
